requirements.txt
asgiref==3.8.1
Django==5.1.3
django-ckeditor==6.7.1
django-cors-headers==4.6.0
django-js-asset==2.2.0
django-taggit==5.0.0
django-taggit-serializer==0.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
mysqlclient==2.1.1
pillow==10.3.0
PyJWT==2.10.0
pytz==2024.2
six==1.17.0
sqlparse==0.5.0
typing_extensions==4.11.0
